pyproject.toml
[tool.poetry]
name = "open-voice-pilot"
version = "0.1.0"
description = ""
authors = ["Bailey Chittle <bachittle@gmail.com>"]
readme = "README.md"
packages = [{include = "open_voice_pilot"}]

[tool.poetry.dependencies]
python = "^3.10"
pyaudio = "*"
ffmpeg-python = "*"
matplotlib = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
